+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Security in case of vulnerabilities.
 ### Changed  
 
 ### Fixed  
+- Fixed incoming trust displaying as `None`
 - update exports tab to point at correct page
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-06-17.413...main) for everything awaiting release